text render experiment.
very rough work, proof of concept font rendering, render unicode code points using a truetype font to a texture in OpenGLES2. Works but implementation is very rough Additionaly I think font rendering should not be part of the base graphics package, it should be a separate, dependant package. Pushing this to remote for reference in future work.